fix: 플레이리스트 카드 수정 #63
